Combine inventory and clue data
