+++ Assets/Dummy/Scriptable Objects/Dummy.meta
