+++ Assets/Dummy/Scriptable Objects/Dummy/Item.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c66440dbeca730d4b971fc3f2fe710fa
+guid: e878e8ea300d0404eb89616b750f37c6
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
