+++ Assets/Dummy/Scriptable Objects/Dummy/Item/item-1.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0d2f2fbc4b1444b4e94fd97e4500fd47, type: 3}
+  m_Name: item-1
+  m_EditorClassIdentifier: 
+  ClueID: item-1
+  ClueName: Item A
+  ClueDescription: Item ini dibutuhkan NPC A
+  ClueImage: {fileID: 728017172041052461, guid: f8579b10a8b9a714e8e1abb6e3c165be, type: 3}
