+++ Assets/Dummy/Scriptable Objects/Dummy/Item/item-1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9722116ee3c70694e892260325c1a78e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
