+++ Assets/Dummy/Scriptable Objects/Dummy/Quest.meta
