+++ Assets/Dummy/Scriptable Objects/Dummy/Quest/quest-1-1.asset.meta
