+++ Assets/Dummy/Scriptable Objects/Dummy/Quest/quest-1-2.asset.meta
