+++ Assets/Dummy/Scriptable Objects/Dummy/Quest/quest-2.asset
