+++ Assets/Scripts/Interaction/ClueData.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 namespace TheDuction.Interaction
 {
-    [CreateAssetMenu(fileName = "New Item", menuName = "Inventory/Item")]
+    [CreateAssetMenu(fileName = "New Item", menuName = "Scriptable Objects/Item")]
     public class ClueData : ScriptableObject
     {
         public string ClueID;
