+++ Assets/Scripts/Interaction/ClueInteractable.cs
@@ -4,12 +4,12 @@ namespace TheDuction.Interaction{
     public class ClueInteractable : Interactable {
         [SerializeField] private ClueData _clueData;
         public delegate void ItemAction(ClueData clueData);
-        public static event ItemAction OnItemAction;
+        public static event ItemAction OnItemInteracted;
 
         public override void Interact()
         {
             base.Interact();
-            OnItemAction?.Invoke(_clueData);
+            OnItemInteracted?.Invoke(_clueData);
             // TODO: Set active false after dialogue
             // gameObject.SetActive(false);
         }
